Update all components with new handlers and props
The `Game` component now has no state. The state has been
lifted up to the `Tot` component. The `Game` component
now takes three props: `squares`, `onClick`, and `id`.
The `Tot` component now handles the all the valid and invalid
plays; is responsible for updating and maintainig the state
of the game.

* Yet to fix: The onClick hnadler for the `Tot` component.